Duplicate usePlaceBid hook in two files
