+++ frontend/afristore-app/src/__tests__/listings.test.tsx
@@ -20,6 +20,9 @@ jest.mock('@/lib/contract', () => ({
 
 jest.mock('@/hooks/useMarketplace', () => ({
   useBuyArtwork: () => ({ buy: jest.fn(), isBuying: false, error: null }),
+}))
+
+jest.mock('@/hooks/usePlaceBid', () => ({
   usePlaceBid: () => ({ bid: jest.fn(), isBidding: false, error: null }),
 }))
 
@@ -34,12 +37,12 @@ jest.mock('@/hooks/useUserActivity', () => ({
 describe('Regression Test: Invalid Listing IDs', () => {
   it('renders Artwork Not Found state for invalid IDs', async () => {
     render(<ListingDetailPage />)
-    
+
     // Wait for the async loading to finish and show error
     await waitFor(() => {
       expect(screen.getByRole('heading', { name: /Artwork Not Found/i })).toBeInTheDocument()
     })
-    
+
     expect(screen.getByRole('heading', { name: /Artwork Not Found/i })).toBeInTheDocument()
     expect(screen.getByText(/Artwork not found/i, { selector: 'p' })).toBeInTheDocument()
     expect(screen.getByRole('button', { name: /Return to Marketplace/i })).toBeInTheDocument()
