+++ frontend/afristore-app/src/__tests__/useAuctions.test.tsx
@@ -45,9 +45,9 @@ import {
   useAuctions,
   useArtistAuctions,
   useCreateAuction,
-  usePlaceBid,
   useFinalizeAuction,
 } from '@/hooks/useAuctions';
+import { usePlaceBid } from '@/hooks/usePlaceBid';
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
