+++ frontend/afristore-app/src/hooks/useAuctions.ts
@@ -167,37 +167,6 @@ export function useCreateAuction(creatorPublicKey: string | null) {
   return { create, isCreating, progress, error };
 }
 
-// ── usePlaceBid ──────────────────────────────────────────────
-
-export function usePlaceBid(bidderPublicKey: string | null) {
-  const [isBidding, setIsBidding] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  useTransientErrorToast(error);
-
-  const bid = useCallback(
-    async (auctionId: number, amountXlm: number): Promise<boolean> => {
-      if (!bidderPublicKey) {
-        setError("Wallet not connected");
-        return false;
-      }
-      setIsBidding(true);
-      setError(null);
-      try {
-        await placeBid(bidderPublicKey, auctionId, amountXlm);
-        return true;
-      } catch (err: unknown) {
-        setError(getReadableErrorMessage(err, "Failed to place bid"));
-        return false;
-      } finally {
-        setIsBidding(false);
-      }
-    },
-    [bidderPublicKey]
-  );
-
-  return { bid, isBidding, error };
-}
-
 // ── useFinalizeAuction ───────────────────────────────────────
 
 export function useFinalizeAuction(callerPublicKey: string | null) {
